+++ client/src/actions/userActions.js
@@ -1,0 +1,26 @@
+import axios from 'axios';
+
+export const ADD_USER = 'ADD_USER';
+export const GET_USER = 'GET_USER';
+
+
+const URL = 'https://donteatthat.herokuapp.com';
+
+
+export const getUser = () => dispatch => {
+    let firebaseid = localStorage.getItem('uid');
+    axios
+      .get(`${URL}/api/users/one/${firebaseid}`)
+      .then(res => dispatch({ type: GET_USER, payload: res.data }))
+      .catch(err => dispatch({ type: ERROR, payload: err }));
+  };
+  
+  export const addUser = firebaseid => dispatch => {
+    axios
+      .post(`${URL}/api/users/create`, { firebaseid })
+      .then(res =>
+        dispatch({ type: ADD_USER, payload: { id: res.data, firebaseid } })
+      )
+      .catch(err => dispatch({ type: ERROR, payload: err }));
+  };
+
